+++ packages/jest-core/src/__tests__/getNoTestsFoundMessage.test.ts
@@ -18,7 +18,7 @@ describe('getNoTestsFoundMessage', () => {
   function createGlobalConfig(options?: Partial<Config.GlobalConfig>) {
     return makeGlobalConfig({
       rootDir: '/root/dir',
-      testPathPattern: '/path/pattern',
+      testPathPatterns: ['/path/pattern'],
       ...options,
     });
   }