+++ packages/jest-core/src/lib/updateGlobalConfig.ts
@@ -6,7 +6,7 @@
  */
 
 import type {Config} from '@jest/types';
-import {replacePathSepForRegex} from 'jest-regex-util';
+import {TestPathPatterns} from 'jest-util';
 import type {AllowedConfigOptions} from 'jest-watcher';
 
 type ExtraConfigOptions = Partial<
@@ -31,15 +31,14 @@ export default function updateGlobalConfig(
     newConfig.testNamePattern = options.testNamePattern || '';
   }
 
-  if (options.testPathPattern !== undefined) {
-    newConfig.testPathPattern =
-      replacePathSepForRegex(options.testPathPattern) || '';
+  if (options.testPathPatterns !== undefined) {
+    newConfig.testPathPatterns = options.testPathPatterns;
   }
 
   newConfig.onlyChanged =
     !newConfig.watchAll &&
     !newConfig.testNamePattern &&
-    !newConfig.testPathPattern;
+    !TestPathPatterns.fromGlobalConfig(newConfig).isSet();
 
   if (typeof options.bail === 'boolean') {
     newConfig.bail = options.bail ? 1 : 0;