+++ packages/jest-types/src/Circus.ts
@@ -203,6 +203,11 @@ export type TestResult = {
   duration?: number | null;
   errors: Array<FormattedError>;
   errorsDetailed: Array<MatcherResults | unknown>;
+  /**
+   * Whether [`test.failing()`](https://jestjs.io/docs/api#testfailingname-fn-timeout)
+   * was used.
+   */
+  failing?: boolean;
   invocations: number;
   status: TestStatus;
   location?: {column: number; line: number} | null;